Fix issue when point is in org src block header
When point is in a org src block header, eg. `#+begin_src` or `#+header:` line, automatic fill will cause the cursor moving to the beginning of src block content and insert a whitespace there. This is unwanted behavior and it makes inserting header args impossible. So we disable filling when point is in such lines.